requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyJWT==2.8.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
requests==2.32.4
